Initial live transcoding implementation
Inspired by the implementation in StashApp, this implements live transcoding as a HLS VoD playlist. We pre-generate the playlist with the expected segments, then dynamically manage the transcoder processes based on the actual media segments requested by clients.

This is a big improvement over our implementation in the original viewer, as it properly handles seeking without requiring custom UI, works in Safari, starts playing much faster, and doesn't require a non-web-standard container format.

There's a fair bit of indirection here with the web request dispatching a queue job, then the queue job starting an artisan command, which then manages the transcoding. This means that all communication has to go via the cache and error handling is a nightmare. However, it keeps the ffmpeg processes to the background queue cgroup rather than the web server's, and it avoids using up critical queue workers with long-running jobs.